+++ src/main/java/dev/kir/sync/api/networking/ShellUpdatePacket.java
@@ -12,15 +12,20 @@ import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtSizeTracker;
 import net.minecraft.network.PacketByteBuf;
+import net.minecraft.network.RegistryByteBuf;
+import net.minecraft.network.codec.PacketCodec;
+import net.minecraft.network.packet.CustomPayload;
 import net.minecraft.util.Identifier;
 
 import java.util.Collection;
 import java.util.List;
 
-public class ShellUpdatePacket implements ClientPlayerPacket {
+public class ShellUpdatePacket implements CustomPayload {
     private Identifier worldId;
     private boolean isArtificial;
     private Collection<ShellState> states;
+    public static final CustomPayload.Id<ShellUpdatePacket> ID = new CustomPayload.Id<>(Sync.locate("packet.shell.update"));
+    public static final PacketCodec<RegistryByteBuf, ShellUpdatePacket> CODEC = PacketCodec.of(ShellUpdatePacket::write, ShellUpdatePacket::new);
 
     public ShellUpdatePacket(Identifier worldId, boolean isArtificial, Collection<ShellState> states) {
         this.worldId = worldId;
@@ -28,12 +33,15 @@ public class ShellUpdatePacket implements ClientPlayerPacket {
         this.states = states == null ? List.of() : states;
     }
 
-    @Override
-    public Identifier getId() {
-        return Sync.locate("packet.shell.update");
+    public ShellUpdatePacket(PacketByteBuf byteBuf) {
+        read(byteBuf);
     }
 
     @Override
+    public Id<ShellUpdatePacket> getId() {
+        return ID;
+    }
+
     public void write(PacketByteBuf buffer) {
         buffer.writeIdentifier(this.worldId);
         buffer.writeBoolean(this.isArtificial);
@@ -41,19 +49,16 @@ public class ShellUpdatePacket implements ClientPlayerPacket {
         this.states.forEach(x -> buffer.writeNbt(x.writeNbt(new NbtCompound())));
     }
 
-    @Override
     public void read(PacketByteBuf buffer) {
         this.worldId = buffer.readIdentifier();
         this.isArtificial = buffer.readBoolean();
         this.states = buffer.readList(subBuffer -> ShellState.fromNbt((NbtCompound) subBuffer.readNbt(NbtSizeTracker.ofUnlimitedBytes())));
     }
 
-    @Override
     public Identifier getTargetWorldId() {
         return this.worldId;
     }
 
-    @Override
     @Environment(EnvType.CLIENT)
     public void execute(MinecraftClient client, ClientPlayerEntity player, ClientPlayNetworkHandler handler, PacketSender responseSender) {
         Shell shell = (Shell)player;
